+++ HISTORY.md
@@ -1,3 +1,7 @@
+# 1.3.0 (NEXT)
+* [BREAK] Remove `visitor` related code, use the new `metadata` field instead
+* Add event emitter for tracking any connection
+
 # 1.2.9 (2016-09-09)
 * Fix #16; Prevent error when proccess exit
 * Fix ESLint errors